refactor(crypto): split cipher.ts into per-concern modules (952 → 542 + 412)
Same per-concern split pattern as recent process / tls / http2 /
canvas2d-core / util / webgl2 PRs. The AES math and the
algorithm/buffer/PKCS#7 helpers move out of cipher.ts into two
focused modules; the three cipher classes stay together (they're
tightly coupled: Decipher inherits from CipherBase like Cipher
does, and the GCM authenticated-encryption path threads through
both).

Layout:

  src/cipher.ts          (542 LoC) — CipherBase + Cipher + Decipher
                                      + createCipher / createCipheriv /
                                      createDecipher / createDecipheriv /
                                      getCiphers factories
  src/aes-primitives.ts  (315 LoC) — SBOX / INV_SBOX / RCON constants,
                                      gmul (GF(2^8)), keyExpansion(),
                                      aesEncryptBlock / aesDecryptBlock,
                                      incrementCounter / gcmIncrementCounter,
                                      gfMul (GF(2^128) — module-private),
                                      ghash() per NIST SP 800-38D
  src/cipher-utils.ts    ( 97 LoC) — interface AlgorithmInfo,
                                      parseAlgorithm, toBuffer,
                                      encodeOutput, incompleteUtf8Tail,
                                      pkcs7Pad / pkcs7Unpad

Public surface preserved verbatim — no module-level exports changed.

572 GJS crypto tests still green; tsc + build clean.

STATUS.md: new Completed entry.
