+++ STATUS.md
@@ -398,6 +398,8 @@ Not yet implemented (but potentially relevant for GJS projects):
 
 ### Completed
 
+- ~~**`@gjsify/crypto` — split `src/cipher.ts` into per-concern modules (952 → 542 + 412)**~~✓ — Same per-concern split pattern. Base `cipher.ts` shrinks from 952 LoC to 542 LoC; the AES math and the algorithm/buffer/PKCS#7 helpers move out into two focused modules. Layout: `aes-primitives.ts` (315 LoC — `SBOX` / `INV_SBOX` / `RCON` constants, the `gmul` GF(2^8) helper, `keyExpansion()`, `aesEncryptBlock()` / `aesDecryptBlock()` block primitives, `incrementCounter()` / `gcmIncrementCounter()` CTR/GCM counter helpers, the GF(2^128) `gfMul` private helper, and the `ghash()` GHASH function per NIST SP 800-38D), `cipher-utils.ts` (97 LoC — `interface AlgorithmInfo`, `parseAlgorithm()`, `toBuffer()`, `encodeOutput()`, `incompleteUtf8Tail()`, `pkcs7Pad()` / `pkcs7Unpad()`). Base file retains `CipherBase`, `Cipher`, `Decipher` classes + the `createCipher` / `createCipheriv` / `createDecipher` / `createDecipheriv` / `getCiphers` factories — these three classes are tightly coupled (Decipher inherits CipherBase like Cipher does, and the GCM authenticated-encryption path threads through both), so they stay together. Public surface preserved verbatim; no module exports changed. 572 GJS crypto tests still green; tsc + build clean. Same approach as the `@gjsify/util` split (PR #268): the extracted modules are pure utility code with stable surface that's easier to spot-check and reuse (e.g. a future hand-rolled GCM authentication-tag verification helper for streaming uploads could import `ghash` directly without dragging in the whole Cipher class).
+
 - ~~**`gjsify install --backend=native|npm` flag — closes the long-standing Phase-4 follow-up**~~✓ — Closes the "🔲 `gjsify install <pkg>` user-facing CLI" Open TODO under "Phase 4: gjsify install / dlx native backend". The TODO entry was stale: the native backend has been the default since the `install-backend.ts` flip much earlier, but the only way to opt back into `npm install` was the undocumented `GJSIFY_INSTALL_BACKEND=npm` env var. New `--backend native|npm` flag on `gjsify install` exposes that toggle as a first-class CLI argument; precedence is explicit `--backend` > env var > `native` default. The env var stays honoured so existing scripts (CI, dotfiles) keep working unchanged. File header + `--help` text updated to document escape-hatch use cases: Yarn PnP repos (the native backend rejects with a clear error), npm lifecycle scripts (deferred behind a separate `--allow-scripts` follow-up that's still open), and npm's per-parent `overrides` shape that the native resolver doesn't yet flatten. No new dependency — the flag just routes the existing `projectInstallViaNpm` / `projectInstallNative` codepaths.
 
 - ~~**`@gjsify/webgl` — extract `WebGL2RenderingContext` object-lifecycle methods (1439 → 1236 + 285)**~~✓ — First slice of the long-running `webgl2-rendering-context.ts` split. Same `install*Methods(proto)` pattern as the parent `webgl-context-base.ts` split, applied to the WebGL2 subclass. Extracts the five object-lifecycle method groups that share a near-identical shape (`createX` / `deleteX` / `isX` / `bindX` plus type-specific param setters/getters): VAO (`createVertexArray` / `deleteVertexArray` / `isVertexArray` / `bindVertexArray`), Query (`createQuery` / `deleteQuery` / `isQuery` / `beginQuery` / `endQuery` / `getQuery` / `getQueryParameter`), Sampler (`createSampler` / `deleteSampler` / `isSampler` / `bindSampler` / `samplerParameteri` / `samplerParameterf` / `getSamplerParameter`), Sync (`fenceSync` / `isSync` / `deleteSync` / `clientWaitSync` / `waitSync` / `getSyncParameter`), Transform Feedback (`createTransformFeedback` / `deleteTransformFeedback` / `isTransformFeedback` / `bindTransformFeedback` / `beginTransformFeedback` / `endTransformFeedback` / `pauseTransformFeedback` / `resumeTransformFeedback` / `transformFeedbackVaryings` / `getTransformFeedbackVarying`). 32 methods totalling 203 LoC moved out of the base file. New module `packages/framework/webgl/src/ts/webgl2-context/object-lifecycle.ts` (285 LoC including the `interface ObjectLifecycleMethods` declaration-merged into `WebGL2RenderingContext` via `declare module`, the methods object, and the `installObjectLifecycleMethods(proto)` function). Base file `webgl2-rendering-context.ts` 1439 → 1236 LoC, calls `installObjectLifecycleMethods(WebGL2RenderingContext.prototype)` after class declaration plus a deliberate side-effect `import './webgl2-context/object-lifecycle.js'` retained in the emitted `.d.ts` so downstream consumers see the augmentations. Same `.d.ts` preservation technique as the canvas2d-core split (PR #262). Unused `WebGLActiveInfo` import dropped from the base. 860 GJS webgl conformance tests still green. Larger remaining method groups (texImage3D family ~190 LoC, uniform family ~110 LoC, clearBuffer family ~110 LoC, instanced drawing + framebuffer-ops ~150 LoC) are left in the base; future PRs can extract them with the same pattern if motivated.
